% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{acc_parity}
\alias{acc_parity}
\title{Accuracy Parity}
\usage{
acc_parity(actuals, predicted, group, cutoff = 0.5, base = NULL)
}
\arguments{
\item{actuals}{Vector of actual target values}

\item{predicted}{Vector of predicted target values}

\item{group}{Sensitive group (binary or factor)}

\item{cutoff}{Cutoff for rounding the probabilities}
}
\value{
Accuracy Parity metric
}
\description{
This function computes the Accuracy Parity metric (Friedler et al. 2018)
}
\examples{
df = fairness::compas
acc_parity(df$label_value, df$score, df$race, "Caucasian")
}
