% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_parity.R
\name{dem_parity}
\alias{dem_parity}
\title{Demographic Parity}
\usage{
dem_parity(predicted, group, cutoff = 0.5, base = NULL)
}
\arguments{
\item{predicted}{Vector of predicted target values}

\item{group}{Sensitive group (binary or factor)}

\item{cutoff}{Cutoff for rounding the probabilities}
}
\value{
DP metric
}
\description{
This function computes the Demographic Parity metric (Calders and Verwer 2010)
}
\examples{
df = fairness::compas
dem_parity(df$score, df$race, "Caucasian")
}
