% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{dis_impact}
\alias{dis_impact}
\title{Disparate Impact}
\usage{
dis_impact(predicted, group, cutoff = 0.5, base = NULL)
}
\arguments{
\item{predicted}{Vector of predicted target values}

\item{group}{Sensitive group (binary or factor)}

\item{cutoff}{Cutoff for rounding the probabilities}
}
\value{
DI metric
}
\description{
This function computes the Disparate Impact metric (Feldman et al. 2015; Zafar et al. 2017)
}
\examples{
df = fairness::compas
dis_impact(df$label_value, df$score, df$race, "Caucasian")
}
