% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairness-package.R
\docType{package}
\name{fairness}
\alias{fairness}
\alias{fairness-package}
\title{fairness: Algorithmic Fairness Metrics}
\description{
The \strong{fairness} package offers ...
}
\details{
\tabular{ll}{
Package: \tab fairness\cr
Depends: \tab R (>= 3.6.0)\cr
Type: \tab Package\cr
Version:  \tab 1.0.0\cr
Date:  \tab 2019-09-02\cr
License:  \tab MIT\cr
LazyLoad:  \tab Yes
}
}
\seealso{
\url{https://github.com/kozodoi/Fairness}
}
\author{
\itemize{
 \item Nikita Kozodoi \email{nikita.kozodoi@hu-berlin.de}
 \item Tibor V. Varga \email{tirgit@hotmail.com}
}
}
