% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{fnr_parity}
\alias{fnr_parity}
\title{FNR Parity}
\usage{
fnr_parity(actuals, predicted, group, cutoff = 0.5, base = NULL)
}
\arguments{
\item{actuals}{Vector of actual target values}

\item{predicted}{Vector of predicted target values}

\item{group}{Sensitive group (binary or factor)}

\item{cutoff}{Cutoff for rounding the probabilities}
}
\value{
FNR Parity metric
}
\description{
This function computes the False Negative Rate Parity metric (Chouldechova 2017)
}
\examples{
df = fairness::compas
fnr_parity(df$label_value, df$score, df$race, "Caucasian")
}
