% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_parity.R
\name{fpr_parity}
\alias{fpr_parity}
\title{FPR Parity}
\usage{
fpr_parity(actuals, predicted, group, cutoff = 0.5, base = NULL)
}
\arguments{
\item{actuals}{Vector of actual target values}

\item{predicted}{Vector of predicted target values}

\item{group}{Sensitive group (binary or factor)}

\item{cutoff}{Cutoff for rounding the probabilities}
}
\value{
FPR Parity metric
}
\description{
This function computes the False Positive Rate Parity metric (Chouldechova 2017)
}
\examples{
df = fairness::compas
fpr_parity(df$label_value, df$score, df$race, "Caucasian")
}
