% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcc_parity.R
\name{mcc_parity}
\alias{mcc_parity}
\title{Matthews Correlation Coefficient parity}
\usage{
mcc_parity(data, outcome, group, probs = NULL, preds = NULL,
  preds_levels = c("no", "yes"), cutoff = 0.5, base = NULL)
}
\arguments{
\item{data}{The dataframe that contains the necessary columns.}

\item{outcome}{The column name of the actual outcomes.}

\item{group}{Sensitive group to examine.}

\item{probs}{The column name of the predicted probabilities (numeric between 0 - 1). If not defined, argument preds need to be defined.}

\item{preds}{The column name of the predicted outcome (categorical outcome). If not defined, argument probs need to be defined.}

\item{preds_levels}{The desired levels of the predicted outcome (categorical outcome). As these levels are commonly defined as yes/no, the function uses this as default.}

\item{cutoff}{Cutoff to generate predicted outcomes from predicted probabilities. Default set to 0.5.}

\item{base}{Base level for sensitive group comparison}
}
\value{
Matthews Correlation Coefficient parity metrics for all groups. Lower values compared to the reference group mean Matthews Correlation Coefficients in the selected subgroups.
}
\description{
This function computes the Matthews Correlation Coefficient (MCC) parity metric
}
\details{
This function computes the Matthews Correlation Coefficient (MCC) parity metric. In the returned
named vector, the reference group will be assigned 1, while all other groups will be assigned values
according to whether their Matthews Correlation Coefficients are lower or higher compared to the reference group. Lower
Matthews Correlation Coefficients rates will be reflected in numbers lower than 1 in the returned named vector, thus numbers
lower than 1 mean WORSE prediction for the subgroup.
}
\examples{
df <- fairness::compas
mcc_parity(data = df, outcome = df$score, group = df$race, base = "Caucasian")

}
