% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{npv_parity}
\alias{npv_parity}
\title{NPV Parity}
\usage{
npv_parity(actuals, predicted, group, cutoff = 0.5, base = NULL)
}
\arguments{
\item{actuals}{Vector of actual target values}

\item{predicted}{Vector of predicted target values}

\item{group}{Sensitive group (binary or factor)}

\item{cutoff}{Cutoff for rounding the probabilities}
}
\value{
FNR Parity metric
}
\description{
This function computes the Negative Positive Value Parity metric (see Aeuquitas bias audit toolkit)
}
\examples{
df = fairness::compas
npv_parity(df$label_value, df$score, df$race, "Caucasian")
}
