% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppv_parity.R
\name{ppv_parity}
\alias{ppv_parity}
\title{Positive Predictive Value parity}
\usage{
ppv_parity(data, outcome, group, probs = NULL, preds = NULL,
  outcome_levels = c("no", "yes"), cutoff = 0.5, base = NULL)
}
\arguments{
\item{data}{The dataframe that contains the necessary columns.}

\item{outcome}{The column name of the actual outcomes.}

\item{group}{Sensitive group to examine.}

\item{probs}{The column name of the predicted probabilities (numeric between 0 - 1). If not defined, argument preds need to be defined.}

\item{preds}{The column name of the predicted outcome (categorical outcome). If not defined, argument probs need to be defined.}

\item{outcome_levels}{The desired levels of the predicted outcome (categorical outcome). As these levels are commonly defined as yes/no, the function uses this as default.}

\item{cutoff}{Cutoff to generate predicted outcomes from predicted probabilities. Default set to 0.5.}

\item{base}{Base level for sensitive group comparison}
}
\value{
\item{Metric}{Positive Predictive Value parity metrics for all groups. Lower values compared to the reference group mean lower positive predictive values in the selected subgroups}
\item{Metric_plot}{Bar plot of Positive Predictive Value metric}
\item{Probability_plot}{Density plot of predicted probabilities per subgroup. Only plotted if probabilities are defined}
}
\description{
This function computes the Positive Predictive Value (PPV) parity metric
}
\details{
This function computes the Positive Predictive Value (PPV) parity metric as described by the Aeuquitas bias toolkit.
Positive Predictive Values (otherwise known as precision) are calculated
by the division of true positives with all predicted positives. In the returned
named vector, the reference group will be assigned 1, while all other groups will be assigned values
according to whether their positive predictive values are lower or higher compared to the reference group. Lower
positive predictive values will be reflected in numbers lower than 1 in the returned named vector, thus numbers
lower than 1 mean WORSE prediction for the subgroup.
}
\examples{
df <- fairness::compas
ppv_parity(data = df, outcome = df$score, group = df$race, base = "Caucasian")

}
