% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_parity.R
\name{pr_parity}
\alias{pr_parity}
\title{Predictive Rate Parity}
\usage{
pr_parity(actuals, predicted, group, cutoff = 0.5, base = NULL)
}
\arguments{
\item{actuals}{Vector of actual target values}

\item{predicted}{Vector of predicted target values}

\item{group}{Sensitive group (binary or factor)}

\item{cutoff}{Cutoff for rounding the probabilities}
}
\value{
Predictive Rate Parity metric
}
\description{
This function computes the Predictive Rate Parity
}
\examples{
df = fairness::compas
pr_parity(df$label_value, df$score, df$race, "Caucasian")
}
