% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_parity.R
\name{roc_parity}
\alias{roc_parity}
\title{ROC AUC parity}
\usage{
roc_parity(data, outcome, group, probs, outcome_levels = c("no", "yes"),
  base = NULL)
}
\arguments{
\item{data}{The dataframe that contains the necessary columns.}

\item{outcome}{The column name of the actual outcomes.}

\item{group}{Sensitive group to examine.}

\item{probs}{The column name of the predicted probabilities (numeric between 0 - 1).}

\item{outcome_levels}{The desired levels of the predicted outcome (categorical outcome). As these levels are commonly defined as yes/no, the function uses this as default.}

\item{base}{Base level for sensitive group comparison}
}
\value{
\item{Metric}{ROC AUC values for all groups. Lower values compared to the reference group mean lower ROC AUC values in the selected subgroups}
\item{ROCAUC_overlap}{The value of overlapping ROC AUC values}
\item{Metric_plot}{Bar plot of ROC AUC metric}
\item{Probability_plot}{Density plot of predicted probabilities per subgroup}
\item{ROCAUC_plot}{ROC plots for all subgroups}
}
\description{
This function computes the Accuracy parity metric
}
\details{
This function computes the ROC AUC values for each subgroup. In the returned
named vector, the reference group will be assigned 1, while all other groups will be assigned values
according to whether their ROC AUC values are lower or higher compared to the reference group. Lower
ROC AUC will be reflected in numbers lower than 1 in the returned named vector, thus numbers
lower than 1 mean WORSE prediction for the subgroup.
}
\examples{
df <- fairness::compas
roc_parity(data = df, outcome = df$score, group = df$race, base = "Caucasian")

}
