% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sens_parity.R
\name{sens_parity}
\alias{sens_parity}
\title{Sensitivity parity}
\usage{
sens_parity(data, outcome, group, probs = NULL, preds = NULL,
  preds_levels = c("no", "yes"), cutoff = 0.5, base = NULL)
}
\arguments{
\item{data}{The dataframe that contains the necessary columns.}

\item{outcome}{The column name of the actual outcomes.}

\item{group}{Sensitive group to examine.}

\item{probs}{The column name of the predicted probabilities (numeric between 0 - 1). If not defined, argument preds need to be defined.}

\item{preds}{The column name of the predicted outcome (categorical outcome). If not defined, argument probs need to be defined.}

\item{preds_levels}{The desired levels of the predicted outcome (categorical outcome). As these levels are commonly defined as yes/no, the function uses this as default.}

\item{cutoff}{Cutoff to generate predicted outcomes from predicted probabilities. Default set to 0.5.}

\item{base}{Base level for sensitive group comparison}
}
\value{
Sensitivity parity metrics for all groups. Lower values compared to the reference group mean lower sensitivities in the selected subgroups.
}
\description{
This function computes the Sensitivity parity metric
}
\details{
This function computes the Sensitivity parity metric. Sensitivities are calculated
by the division of true positives with all positives (irrespective of predicted values). In the returned
named vector, the reference group will be assigned 1, while all other groups will be assigned values
according to whether their sensitivities are lower or higher compared to the reference group. Lower
sensitivities will be reflected in numbers lower than 1 in the returned named vector, thus numbers
lower than 1 mean WORSE prediction for the subgroup.
}
\examples{
df <- fairness::compas
sens_parity(data = df, outcome = df$score, group = df$race, base = "Caucasian")

}
