% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dem_parity.R
\name{dem_parity}
\alias{dem_parity}
\title{Demographic parity}
\usage{
dem_parity(
  data,
  outcome,
  group,
  probs = NULL,
  preds = NULL,
  preds_levels = NULL,
  outcome_base = NULL,
  cutoff = 0.5,
  base = NULL
)
}
\arguments{
\item{data}{The dataframe that contains the necessary columns.}

\item{outcome}{The column name of the actual outcomes.}

\item{group}{Sensitive group to examine.}

\item{probs}{The column name or vector of the predicted probabilities (numeric between 0 - 1). If not defined, argument preds needs to be defined.}

\item{preds}{The column name or vector of the predicted binary outcome (0 or 1). If not defined, argument probs needs to be defined.}

\item{preds_levels}{The desired levels of the predicted binary outcome. If not defined, levels of the outcome variable are used.}

\item{outcome_base}{Base level for the target variable used to compute fairness metrics. Default is the first level of the outcome variable.}

\item{cutoff}{Cutoff to generate predicted outcomes from predicted probabilities. Default set to 0.5.}

\item{base}{Base level for sensitive group comparison}
}
\value{
\item{Metric}{Absolute number of positive classifications for all groups and metrics standardized for the base group (demographic parity metric). Lower values compared to the reference group mean lower number of positively predicted observations in the selected subgroups}
\item{Metric_plot}{Bar plot of Demographic parity metric}
\item{Probability_plot}{Density plot of predicted probabilities per subgroup. Only plotted if probabilities are defined}
}
\description{
This function computes the Demographic parity metric
}
\details{
This function computes the Demographic parity metric (also known as Statistical Parity, Equal Parity,
Equal Acceptance Rate or Independence) as described by Calders and Verwer 2010. Demographic parity is calculated
based on the comparison of the absolute number of all positively classified individuals in all subgroups of the data. In the returned
named vector, the reference group will be assigned 1, while all other groups will be assigned values
according to whether their proportion of positively predicted observations are lower or higher compared to the reference group. Lower
proportions will be reflected in numbers lower than 1 in the returned named vector.
}
\examples{
data(compas)
dem_parity(data = compas, outcome = 'Two_yr_Recidivism', group = 'ethnicity',
probs = 'probability', preds = NULL,
cutoff = 0.4, base = 'Caucasian')
dem_parity(data = compas, outcome = 'Two_yr_Recidivism', group = 'ethnicity',
probs = NULL, preds = 'predicted',
cutoff = 0.5, base = 'Hispanic')

}
