% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairness-package.R
\docType{package}
\name{fairness}
\alias{fairness}
\alias{fairness-package}
\title{fairness: Algorithmic Fairness Metrics}
\description{
The \strong{fairness} package offers various metrics of algorithmic fairness. Fairness in machine learning is an emerging
topic with the overarching aim to critically assess algorithms (predictive and classification models) whether
their results reinforce existing social biases. While unfair algorithms can propagate such biases and offer
prediction or classification results with a disparate impact on various sensitive subgroups of populations (defined
by sex, gender, ethnicity, religion, income, socioeconomic status, physical or mental disabilities), fair algorithms possess
the underlying foundation that these groups should be treated similarly / should have similar outcomes. The fairness
R package offers the calculation and comparisons of commonly and less commonly used fairness metrics in population
subgroups. The package also offers convenient visualizations to help understand fairness metrics.
}
\details{
\tabular{ll}{
Package: \tab fairness\cr
Depends: \tab R (>= 3.5.0)\cr
Type: \tab Package\cr
Version:  \tab 1.0.1\cr
Date:  \tab 2019-09-19\cr
License:  \tab MIT\cr
LazyLoad:  \tab Yes
}
}
\seealso{
\url{https://github.com/kozodoi/Fairness}
}
\author{
\itemize{
 \item Nikita Kozodoi \email{nikita.kozodoi@hu-berlin.de}
 \item Tibor V. Varga \email{tirgit@hotmail.com}
}
}
