% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppv_parity.R
\name{ppv_parity}
\alias{ppv_parity}
\title{PPV Parity}
\usage{
ppv_parity(actuals, predicted, group, cutoff = 0.5, base = NULL)
}
\arguments{
\item{actuals}{Vector of actual target values}

\item{predicted}{Vector of predicted target values}

\item{group}{Sensitive group (binary or factor)}

\item{cutoff}{Cutoff for rounding the probabilities}
}
\value{
FNR Parity metric
}
\description{
This function computes the Positive Predicted Value Parity metric (see Aeuquitas bias audit toolkit)
}
\examples{
df = fairness::compas
ppv_pariy(df$label_value, df$score, df$race, "Caucasian")
}
